+++ Cyotek.Windows.Forms.TabList/VisualStudioTabListRenderer.cs
@@ -1,10 +1,9 @@
 using System.Drawing;
-using System.Windows.Forms;
 
 namespace Cyotek.Windows.Forms
 {
   // Cyotek TabList
-  // Copyright (c) 2012-2013 Cyotek.
+  // Copyright (c) 2012-2020 Cyotek.
   // https://www.cyotek.com
   // https://www.cyotek.com/blog/tag/tablist
 
@@ -36,43 +35,19 @@ namespace Cyotek.Windows.Forms
     /// <seealso cref="M:Cyotek.Windows.Forms.TabListPageRenderer.RenderHeader(Graphics,TabListPage,TabListPageState)"/>
     public override void RenderHeader(Graphics g, TabListPage page, TabListPageState state)
     {
-      Color fillColor;
-      Color textColor;
-      Rectangle fillBounds;
-      Rectangle textRectangle;
-      TextFormatFlags flags;
       int arrowSize;
+      Rectangle fillBounds;
+      Rectangle textBounds;
 
       arrowSize = 6;
       fillBounds = page.HeaderBounds;
       fillBounds.Width -= arrowSize;
-      textRectangle = Rectangle.Inflate(fillBounds, -4, -4);
+      textBounds = Rectangle.Inflate(fillBounds, -4, -4);
 
-      // define the most appropriate colors
-      if ((state & TabListPageState.Selected) == TabListPageState.Selected)
-      {
-        fillColor = SystemColors.Highlight;
-        textColor = SystemColors.HighlightText;
-      }
-      else if ((state & TabListPageState.HotLight) == TabListPageState.HotLight)
-      {
-        fillColor = SystemColors.ControlLightLight;
-        textColor = page.Owner.ForeColor;
-      }
-      else
-      {
-        fillColor = Color.Transparent;
-        textColor = page.Owner.ForeColor;
-      }
-
-      // fill the background
-      if (fillColor.A > 0)
-      {
-        using (Brush brush = new SolidBrush(fillColor))
-        {
-          g.FillRectangle(brush, fillBounds);
-        }
-      }
+      RenderSupport.DefineColors(page, state, out Color fillColor, out Color textColor, SystemColors.ControlLightLight);
+      RenderSupport.DrawBackground(g, fillBounds, fillColor);
+      RenderSupport.DrawText(g, page, textBounds, textColor, fillColor);
+      RenderSupport.DrawFocus(g, page, state, textBounds, fitToText: true);
 
       // draw the selection arrow
       if ((state & TabListPageState.Selected) == TabListPageState.Selected)
@@ -83,12 +58,12 @@ namespace Cyotek.Windows.Forms
         Point point2;
         Point point3;
 
-        y = fillBounds.Top + (fillBounds.Height - arrowSize * 2) / 2;
+        y = fillBounds.Top + ((fillBounds.Height - (arrowSize * 2)) / 2);
         x = fillBounds.Right;
 
         point1 = new Point(x, y);
         point2 = new Point(x + arrowSize, y + arrowSize);
-        point3 = new Point(x, y + arrowSize * 2);
+        point3 = new Point(x, y + (arrowSize * 2));
 
         using (Brush brush = new SolidBrush(fillColor))
         {
@@ -99,22 +74,6 @@ namespace Cyotek.Windows.Forms
                                  point3
                                });
         }
-      }
-
-      // draw the text
-      flags = TextFormatFlags.VerticalCenter | TextFormatFlags.Left | TextFormatFlags.NoPrefix | TextFormatFlags.SingleLine | TextFormatFlags.WordEllipsis;
-      TextRenderer.DrawText(g, page.Text, page.Font, textRectangle, textColor, fillColor, flags);
-
-      // focus
-      if ((state & TabListPageState.Focused) == TabListPageState.Focused)
-      {
-        SizeF textSize;
-        int offset;
-
-        textSize = TextRenderer.MeasureText(g, page.Text, page.Font, textRectangle.Size, flags);
-        offset = 2;
-
-        NativeMethods.DrawFocusRectangle(g, textRectangle.X, textRectangle.Y, (int)textSize.Width + offset, (int)textSize.Height + offset);
       }
     }